Change refresh to change current account instead of auth token.
After refresh in the previous version only the authorization token
is being updated. That means that any other information, especially
interesting expiration date, is not updated.

In real-world situation that means that even after refresh was happened
AuthzModule.isAuthorized will return negative anyway, because expiration date
is not updated and stays old, i. e. token is considered expired
even if it was refreshed just fine and can be applied to requests.